move validate functions
