+++ nowcasting_dataset/dataset/validate.py
@@ -3,6 +3,7 @@ A class to validate the prepare ml dataset
 """
 from typing import Union
 
+import numpy as np
 import pandas as pd
 import torch
 
@@ -11,9 +12,23 @@ from nowcasting_dataset.consts import (
     GSP_DATETIME_INDEX,
     DEFAULT_N_GSP_PER_EXAMPLE,
     DEFAULT_N_PV_SYSTEMS_PER_EXAMPLE,
+    GSP_ID,
+    GSP_YIELD,
+    GSP_X_COORDS,
+    GSP_Y_COORDS,
+    OBJECT_AT_CENTER,
+    PV_SYSTEM_ID,
+    PV_YIELD,
+    PV_SYSTEM_X_COORDS,
+    PV_SYSTEM_Y_COORDS,
+    PV_SYSTEM_ROW_NUMBER,
+    PV_AZIMUTH_ANGLE,
+    PV_ELEVATION_ANGLE,
+    DATETIME_FEATURE_NAMES,
 )
 from nowcasting_dataset.dataset import example
 from nowcasting_dataset.dataset.datasets import NetCDFDataset, logger
+from nowcasting_dataset.dataset.example import Example
 
 
 class ValidatorDataset:
@@ -71,7 +86,7 @@ class ValidatorDataset:
         Returns: list of days that the batch has data for
 
         """
-        example.validate_batch_from_configuration(batch, configuration=self.configuration)
+        validate_batch_from_configuration(batch, configuration=self.configuration)
 
         if type(batch[GSP_DATETIME_INDEX]) == torch.Tensor:
             batch[GSP_DATETIME_INDEX] = batch[GSP_DATETIME_INDEX].detach().numpy()
@@ -171,3 +186,157 @@ class FakeDataset(torch.utils.data.Dataset):
         x["gsp_yield"] = torch.clip(x["gsp_yield"], min=0, max=1)
 
         return x
+
+
+def validate_example(
+    data: Example,
+    seq_len_30_minutes: int,
+    seq_len_5_minutes: int,
+    sat_image_size: int = 64,
+    n_sat_channels: int = 1,
+    nwp_image_size: int = 0,
+    n_nwp_channels: int = 1,
+    n_pv_systems_per_example: int = DEFAULT_N_PV_SYSTEMS_PER_EXAMPLE,
+    n_gsp_per_example: int = DEFAULT_N_GSP_PER_EXAMPLE,
+    batch: bool = False,
+):
+    """
+    Validate the size and shape of the data
+    Args:
+        data: Typed dictionary of the data
+        seq_len_30_minutes: the length of the sequence for 30 minutely data
+        seq_len_5_minutes: the length of the sequence for 5 minutely data
+        sat_image_size: the satellite image size
+        n_sat_channels: the number of satellite channgles
+        nwp_image_size: the nwp image size
+        n_nwp_channels: the number of nwp channels
+        n_pv_systems_per_example: the number pv systems with nan padding
+        n_gsp_per_example: the number gsp systems with nan padding
+        batch: if this example class is a batch or not
+    """
+
+    n_gsp_id = data[GSP_ID].shape[-1]
+    assert (
+        n_gsp_id == n_gsp_per_example
+    ), f"gsp_is is len {n_gsp_id}, but should be {n_gsp_per_example}"
+    assert data[GSP_YIELD].shape[-2:] == (
+        seq_len_30_minutes,
+        n_gsp_id,
+    ), f"gsp_yield is size {data[GSP_YIELD].shape}, but should be {(seq_len_30_minutes, n_gsp_system_id)}"
+    assert data[GSP_X_COORDS].shape[-1] == n_gsp_id
+    assert data[GSP_Y_COORDS].shape[-1] == n_gsp_id
+    assert data[GSP_DATETIME_INDEX].shape[-1] == seq_len_30_minutes
+
+    # check the GSP data is between 0 and 1
+    assert (
+        np.nanmax(data[GSP_YIELD]) <= 1.0
+    ), f"Maximum GSP value is {np.nanmax(data[GSP_YIELD])} but it should be <= 1"
+    assert (
+        np.nanmin(data[GSP_YIELD]) >= 0.0
+    ), f"Maximum GSP value is {np.nanmin(data[GSP_YIELD])} but it should be >= 0"
+
+    if OBJECT_AT_CENTER in data.keys():
+        assert data[OBJECT_AT_CENTER] == "gsp"
+
+    if not batch:
+        # add an extract dimension so that its similar to batch data
+        data["x_meters_center"] = np.expand_dims(data["x_meters_center"], axis=0)
+        data["y_meters_center"] = np.expand_dims(data["y_meters_center"], axis=0)
+
+    # loop over batch
+    for d in data["x_meters_center"]:
+        assert type(d) in [
+            np.float64,
+            torch.Tensor,
+        ], f"x_meters_center should be np.float64 but is {type(d)}"
+    for d in data["y_meters_center"]:
+        assert type(d) in [
+            np.float64,
+            torch.Tensor,
+        ], f"y_meters_center should be np.float64 but is {type(d)}"
+
+    assert data[PV_SYSTEM_ID].shape[-1] == n_pv_systems_per_example
+    assert data[PV_YIELD].shape[-2:] == (seq_len_5_minutes, n_pv_systems_per_example)
+    assert data[PV_SYSTEM_X_COORDS].shape[-1] == n_pv_systems_per_example
+    assert data[PV_SYSTEM_Y_COORDS].shape[-1] == n_pv_systems_per_example
+
+    if not batch:
+        # add an extract dimension so that its similar to batch data
+        data[PV_SYSTEM_ID] = np.expand_dims(data[PV_SYSTEM_ID], axis=0)
+        data[PV_SYSTEM_ROW_NUMBER] = np.expand_dims(data[PV_SYSTEM_ID], axis=0)
+
+    # loop over batch
+    for i in range(len(data[PV_SYSTEM_ID])):
+        n_pv_systems = (data[PV_SYSTEM_ID][i, ~np.isnan(data[PV_SYSTEM_ID][i])]).shape[-1]
+        n_pv_syetem_row_numbers = (
+            data[PV_SYSTEM_ROW_NUMBER][i, ~np.isnan(data[PV_SYSTEM_ROW_NUMBER][i])]
+        ).shape[-1]
+        assert n_pv_syetem_row_numbers == n_pv_systems, (
+            f"Number of PV systems ({n_pv_systems}) does not match the "
+            f"pv systems row numbers ({n_pv_syetem_row_numbers})"
+        )
+
+        if n_pv_systems > 0:
+            # check the PV data is between 0 and 1
+            assert (
+                np.nanmax(data[PV_YIELD]) <= 1.0
+            ), f"Maximum PV value is {np.nanmax(data[PV_YIELD])} but it should be <= 1"
+            assert (
+                np.nanmin(data[PV_YIELD]) >= 0.0
+            ), f"Maximum PV value is {np.nanmin(data[PV_YIELD])} but it should be <= 1"
+
+    if PV_AZIMUTH_ANGLE in data.keys():
+        assert data[PV_AZIMUTH_ANGLE].shape[-2:] == (seq_len_5_minutes, n_pv_systems_per_example)
+    if PV_AZIMUTH_ANGLE in data.keys():
+        assert data[PV_ELEVATION_ANGLE].shape[-2:] == (seq_len_5_minutes, n_pv_systems_per_example)
+
+    assert data["sat_data"].shape[-4:] == (
+        seq_len_5_minutes,
+        sat_image_size,
+        sat_image_size,
+        n_sat_channels,
+    )
+    assert data["sat_x_coords"].shape[-1] == sat_image_size
+    assert data["sat_y_coords"].shape[-1] == sat_image_size
+    assert data["sat_datetime_index"].shape[-1] == seq_len_5_minutes
+
+    nwp_correct_shape = (
+        n_nwp_channels,
+        seq_len_5_minutes,
+        nwp_image_size,
+        nwp_image_size,
+    )
+    nwp_shape = data["nwp"].shape[-4:]
+    assert (
+        nwp_shape == nwp_correct_shape
+    ), f"NWP shape should be ({nwp_correct_shape}), but instead it is {nwp_shape}"
+    assert data["nwp_x_coords"].shape[-1] == nwp_image_size
+    assert data["nwp_y_coords"].shape[-1] == nwp_image_size
+    assert data["nwp_target_time"].shape[-1] == seq_len_5_minutes
+
+    for feature in DATETIME_FEATURE_NAMES:
+        assert data[feature].shape[-1] == seq_len_5_minutes
+
+
+def validate_batch_from_configuration(data: Example, configuration: Configuration):
+    """
+    Validate data using a configuration
+
+    Args:
+        data: batch of data
+        configuration: confgiruation of the data
+
+    """
+
+    validate_example(
+        data=data,
+        seq_len_30_minutes=configuration.process.seq_len_30_minutes,
+        seq_len_5_minutes=configuration.process.seq_len_5_minutes,
+        sat_image_size=configuration.process.satellite_image_size_pixels,
+        n_sat_channels=len(configuration.process.sat_channels),
+        nwp_image_size=configuration.process.nwp_image_size_pixels,
+        n_nwp_channels=len(configuration.process.nwp_channels),
+        n_pv_systems_per_example=DEFAULT_N_PV_SYSTEMS_PER_EXAMPLE,
+        n_gsp_per_example=DEFAULT_N_GSP_PER_EXAMPLE,
+        batch=True,
+    )
